Fixing bug, rename app, change hierarchy - Bug fix - Tasks were deleted incorrectly. - New name app - Voltage - New icon

+++ app/src/main/java/com/george/android/tasker/ui/tasks/folder/EditFolderTaskBottomSheet.java
@@ -1,5 +1,8 @@
 package com.george.android.tasker.ui.tasks.folder;
 
+import static java.util.Objects.requireNonNull;
+
+import android.app.AlertDialog;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,14 +13,14 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.george.android.tasker.data.tasks.folder.TaskFolder;
-import com.george.android.tasker.data.tasks.folder.TaskFolderAdapter;
+import com.george.android.tasker.R;
+import com.george.android.tasker.data.model.TaskFolder;
+import com.george.android.tasker.ui.adapters.TaskFolderAdapter;
+import com.george.android.tasker.data.viewmodel.TasksFolderViewModel;
 import com.george.android.tasker.databinding.EditFolderTaskBottomSheetBinding;
-import com.george.android.tasker.ui.tasks.task.TasksViewModel;
+import com.george.android.tasker.data.viewmodel.TasksViewModel;
 import com.george.android.tasker.utils.Utils;
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment;
-
-import java.util.Objects;
 
 public class EditFolderTaskBottomSheet extends BottomSheetDialogFragment {
 
@@ -37,7 +40,7 @@ public class EditFolderTaskBottomSheet extends BottomSheetDialogFragment {
         View root = binding.getRoot();
         Utils utils = new Utils();
 
-        Objects.requireNonNull(getDialog()).getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE);
+        requireNonNull(getDialog()).getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_RESIZE);
         utils.showSoftKeyboard(binding.nameFolderTextLayout, EditFolderTaskBottomSheet.this.requireActivity());
 
         tasksFolderViewModel = new ViewModelProvider(this).get(TasksFolderViewModel.class);
@@ -51,18 +54,26 @@ public class EditFolderTaskBottomSheet extends BottomSheetDialogFragment {
         position = getArguments().getInt("position");
         name = getArguments().getString("name");
 
-        Objects.requireNonNull(binding.nameFolderTextLayout.getEditText()).setText(name);
+        requireNonNull(binding.nameFolderTextLayout.getEditText()).setText(name);
 
-        binding.deleteFolderTaskBtn.setOnClickListener(v -> {
-            //Удаляем папку
-            TaskFolder taskFolder = taskFolderAdapter.getTaskFolderAt(position);
-            tasksFolderViewModel.delete(taskFolder);
+        binding.deleteFolderTaskBtn.setOnClickListener(v ->
+                new AlertDialog.Builder(EditFolderTaskBottomSheet.this.requireActivity())
+                        .setTitle(R.string.warning_text)
+                        .setMessage(R.string.conform_delete_folder_tasks)
+                        .setPositiveButton("Да", (dialog, which) -> {
+                            //Удаляем папку
+                            TaskFolder taskFolder = taskFolderAdapter.getTaskFolderAt(position);
+                            tasksFolderViewModel.delete(taskFolder);
 
-            //Удаляем все элементы из папки
-            tasksViewModel.deleteTasksFolder(folderId);
+                            //Удаляем все элементы из папки
+                            tasksViewModel.deleteTasksFolder(folderId);
 
-            dismiss();
-        });
+                            dialog.dismiss();
+                            dismiss();
+                        })
+                        .setNegativeButton("Отмена", (dialog, which) -> dialog.dismiss())
+                        .create()
+                        .show());
 
         binding.saveFolderTaskBtn.setOnClickListener(v -> {
             String name = binding.nameFolderTextLayout.getEditText().getText().toString();
@@ -73,6 +84,12 @@ public class EditFolderTaskBottomSheet extends BottomSheetDialogFragment {
         });
 
         return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
     }
